refactored code such that it can * be used as a third party package with import event_cnn_minimal.load_model * can be run as an independent script with python -m event_cnn_minimal.load_model -h
