--- base/__init__.py
@@ -1,1 +1,0 @@
-from .base_model import *
