+++ utils/parse_config.py
@@ -3,7 +3,7 @@ from pathlib import Path
 from functools import reduce, partial
 from operator import getitem
 from datetime import datetime
-from utils import read_json, write_json
+from .util import read_json, write_json
 
 
 class ConfigParser:
